move NameStandardizer classes to the gsrs.module.substance.standardizer directory
