+++ gsrs-module-substance-example/src/test/java/example/name/NameUtilitiesTest.java
@@ -6,11 +6,11 @@ import java.util.List;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
-import gsrs.module.substance.utils.FDAFullNameStandardizer;
-import gsrs.module.substance.utils.NameStandardizer;
+import gsrs.module.substance.standardizer.FDAFullNameStandardizer;
+import gsrs.module.substance.standardizer.NameStandardizer;
+import gsrs.module.substance.standardizer.ReplacementNote;
+import gsrs.module.substance.standardizer.ReplacementResult;
 import gsrs.module.substance.utils.NameUtilities;
-import gsrs.module.substance.utils.ReplacementNote;
-import gsrs.module.substance.utils.ReplacementResult;
 import lombok.extern.slf4j.Slf4j;
 
 /**
