+++ gsrs-module-substances-core/src/main/java/ix/ginas/utils/validation/validators/BasicNameValidator.java
@@ -1,7 +1,7 @@
 package ix.ginas.utils.validation.validators;
 
 import gsrs.module.substance.utils.NameUtilities;
-import gsrs.module.substance.utils.ReplacementResult;
+import gsrs.module.substance.standardizer.ReplacementResult;
 import ix.core.validator.GinasProcessingMessage;
 import ix.core.validator.ValidatorCallback;
 import ix.ginas.models.v1.Substance;
