+++ gsrs-module-substances-core/src/main/resources/substances-core.conf
@@ -243,8 +243,8 @@ gsrs.validators.substances = [
 #                                              "newObjClass" = "ix.ginas.models.v1.Substance",
 #                                              "configClass" = "SubstanceValidatorConfig",
 #                                              "parameters"= {
-#                                                    "inPlaceNameStandardizerClass":"gsrs.module.substance.utils.FDAMinimumNameStandardizer",
-#                                                    "fullNameStandardizerClass":"gsrs.module.substance.utils.FDAFullNameStandardizer",
+#                                                    "inPlaceNameStandardizerClass":"gsrs.module.substance.standardizer.FDAMinimumNameStandardizer",
+#                                                    "fullNameStandardizerClass":"gsrs.module.substance.standardizer.FDAFullNameStandardizer",
 #                                                    "BehaviorOnInvalidStdName": "warn"
 #                                              }
 #                                     }
